Add /style-guide
As recommended in the govuk_admin_template readme, mount /style-guide. The
mounted engine requires root_path to be working, so I added a placeholder
root 'application#index' to make it work - this can be removed as soon as we
have a proper root url. I've also renamed the healthcheck_spec to
utility_pages_spec and added a spec for /style-guide