Add 'Sign out' link
Add a sign out link to the navbar header. It also shows the name of the
currently logged in user. Ideally I would like to write a feature spec for this,
but we don't have any controllers on master to test this on and gds-sso is
making it really difficult too [1].

[1] - https://github.com/alphagov/gds-sso/issues/82