Add state to GuideEdition
States will represent the current state of an edition, it can be draft,
published, 2i etc, we will define it later. For now add extra associations on
Guide to fetch the current draft_edition and the currently published_edition
if they are present.

More work needs to be done to ensure there's only one edition for each state or
some other means of deducting how to fetch the right edition for each case.