Fix a failing test
The test had two issues

1. latest_edition was nil

Assigning an array of editions and referring to latest_edition in a callback
did not work as these are two separate relationships. I believe callbacks are
executed in DB transactions, therefore latest_edition was not persisted when
it was being accessed. Assigning latest_edition directly solves this problem.

2. Calls to publishing-api were not stubbed

I've stubbed out the private #publish method on a guide as this test is not
concerned with the publishing process.