Add a very basic preview link
Add a link to the guide form controls that allows a content designer to preview
a version of a guide that is currently saved in draft content store. UI is not
great as it does not explain that content needs to be saved before previewing,
but it works as a quick helper for demos and development purposes.