Replace dashboards#index with guides#index
The dashboard page currently just lists all guides. Move it to the guides
controller and list them in a table. This is in preparation for the basics of
the content publishing workflow. Different kinds of actions will be required
depending on the state of each guide so we'll need to indicate in in the UI
somehow.