Record each draft save/publication as a separate edition
Whenever changes are made, create a new guide edition. We used to do that for
currently published guides, but not for guides in draft state. Recording each
edition save aligns us with the Performance Platform team (responsible for
publishing-api) future plans for versioning. It also gives us better
accountability tracking as author of each change is now recorder.

Note that I opted to "clone" the previous edition before applying attributes
sent from the form. This is to ensure that we will not loose any data if it is
not passed from the form due to any potential changes in future.