Migrate draft saving and publishing to publishing-api V2
Migrate to V2 without changing behaviour or public APIs in the publisher
application. This means that when publishing documents 2 API requests will be
made: one for saving an edition and one for publishing it. We're still working
on defining the publishing workflow so I did not wan to guess what it's going
to be at this stage yet.