Rename GuidePublisher#publish! to GuidePublisher#process
The method name `publish!` is confusing, because it implies guides get
published when the method is invoked. The truth is guides are either saved as
drafts or published. I've gotten rid of an `!` at the method name too, because
the convention in Ruby is to use exclamation mark for methods that are
considered more dangerous than their counterparts without an exclamation mark.
We don't have one.