Spike: Service Manual Section page hierarchy
Since we can't use any existing publishers for section pages and don't want to
build our own just yet, create a rake task that stores and publishes something
we'd expect a Service Manual Section payload to look like. It's just the link
grouping part that's interesting. Note:

- It does not have any data that deals with breadcrumbs, publishing platform
  team is working on it.
- The hierarchy in the YAML file is generated by scraping http://sm-11.herokuapp.com
- It is using links hash with content IDs alongside details -> linked items,
  the former allows the frontend to get all required info about linked items
  and the later is used for grouping the linked items.

The remaining part of this spike that renders this data is here:
https://github.com/alphagov/government-frontend/compare/service_manual_section_spike

An example page of a Service Manual Section page: http://sm-11.herokuapp.com/agile-delivery/