Make sure no work is lost when previewing drafts
The old behaviour of the preview functionality was displaying the latest saved
draft, that means a draft has to be saved before previewing it. This is
problematic because:
- it errors if there are no persisted editions
- it discards any work in progress done on the form page
- it is not clear that the latest saved edition is shown on preview.

By saving the draft before previewing it can solve all of the above problems. I
tried to make it clear what the button does by renaming the button from
"Preview" to "Save Draft and Preview".

I had to write a expect_external_redirect_to "matcher" because Rack::Test does
not support external redirects and raises ActionController::RoutingError. By
rescuing the error and asserting the current URL we can make sure the code is
executing the redirect we're expecting.