Refactor guide_publishing_flow_spec
While working on this file I found it slightly difficult to read the assertions
as they were checking values that are set in test setup methods and are not
obvious when reading individual scenarios. I've refactored the setup methods to
accept more values and modified scenarios to explicitly set the values that are
later used in assertions.