Avoid weird errors when saving new guides
'Save Draft and Preview' button is very likely to be opened in a new tab by the
users. This is problematic because if a new guide is being created, both
'Save Draft' and 'Save Draft and Preview' buttons opened in a new tab would
persist a new record and keep the old form open in the old tab. Saving a record
from that form would attempt to create a duplicate record of the previously
saved one. To avoid this confusing situation, disable the buttons that save
in case a new record is being created.