Simplify draft preview url generation
Instead of providing different subdomains for production and other environments
always use draft-origin. Initially we thought it's better to use
'government-frontend.dev.gov.uk', but it only works with published guides, so
do not do that anymore as it may be confusing.

A developer who wants to preview content on a dev machine has to run the draft
stack locally under draft-origin.dev.gov.uk