Set phase for guide editions to 'alpha' by default
This field was first introduced because it's a required field by default in
govuk-content-schemas. It has now become clear that we'll only use it to
indicate to the front-end application whether it should display an alpha/beta
banner. The banner would always be the same across all Service Manual content
so there is no need to display it in the UI. I've set it to 'alpha' initially,
once we decide to migrate to beta we'll need to do it with all persisted guides
locally as well as migrating published guides via publishing-api.