Rename publisher_* to content_owner_*
We have realised that using publisher to indicate the group responsible for a
piece of content can be ambiguous because within GOV.UK we use the term
"publisher" to name applications that publish content.

Other repos that need this renaming: govuk-content-schemas, government-frontend
We will not need to sync deploys or update because none of those are used in
production.