Show changes since the last published edition in the UI
Content designers can see changes in the current or any other historic version
of an edition compared to the previously published edition. This is helpful
because they do not need to read both versions and compare them manually to see
new additions, deletions and changes.

It currently has no design and therefore the markup is not great either. I will
wait for design before making assumptions on what it's going to look like.