Enable changes tab for new guides
We used to disable the "Changes" link for guides that do not have any
previously published editions. User research shows this is confusing,
especially given the fact that initially all users are going to be new and
they will work on new content. This changes enables the changes tab for all
editions and compares the latest one with an empty edition if there is no
previously published ones, resulting in all fields being "green" additions.