Retain changes in the guide form on failure
I've noticed that after a validation error or an API error is triggered, the
form loads a wrong edition object and thus loses all the changes made by user.
This can make people really frustrated.

This change makes sure we always use the same latest_edition object in the form.
We tried this attempt before and reverted, because using a has_one association
would replace the published edition (published editions can not be modified).

To avoid this situation the code now ensures there is an editable draft edition
just before updating the existing latest_edition. In such a case it overwrites
the latest_edition id parameter that comes from the form to ensure the newly
created draft edition is updated.

A new side effect of this behaviour can be observed when an update to a
currently published version of guide fails. Then we persist a draft clone of
the latest published edition in the local database. Users should not experience
any difference in the UX compared to the previous behaviour.