Default to major update types for new drafts
When a new draft of a published edition is created, default to update_type
'major', then after saving respect what the user has chosen when first saving
the draft.

We want to encourage users to provide change notes for each change.

Some feature specs broke due to the new default, because major updates require
a change note to be present.