Normalise newlines in seed.rb
When importing data from the old service manual guides the newline symbols are
"\n", but when saved again via the application they change to "\r\n". This
causes "compare changes" view to highlight the whole document first time it's
saved. Importing data with "\r\n" as newlines solves this problem.