Disable publishing based on an ENV variable
We intend to deploy the application to production but we are not ready to
publish any documents yet (we need to coordinate old content replacement,
prepare topic pages etc.). Having the application in production will allow
content designers to work on drafts without losing them every night (as in
preview/integration environment).

We intend to set this environment variable in production instances only (not
staging).