Improve diffing UI based on user feedback
The results from the content design survey weren't totally conclusive but the
eventual winner was "+/- at the start of the rows with a strike through for
deletions".

The only point of confusion around this option was the +/- sit a little too
close to the rest of the body copy, and thus looks a bit like the markdown
itself. Increase the margin between the two.