Allow combining filters with free-text search
I've noticed that the application fails with an SQL error if we have both
filters (e.g. State: draft) and free-text keywords in the query. This is caused
by a modified 'from' clause in the free-text search method.
This commit fixes the issue by not modifying the from clause and duplicating
the to_tsquery function call where required.