Improve slug validation
We've received feedback that the form shows the following error:
"Base path is not a valid absolute URL path, Routes is not a valid absolute URL
path" when a user forgets to changed the slug value that is pre-filled with
'/service-manual/' by default. That error comes from the publishing API and is
a bit confusing. Validate it in the application and improve the error messages.