Notify users when there are unsaved changes
Ask users to confirm their navigation away when the guides form has unsaved
changes. It works by recording a checksum of the form values on page load, then
checking if the checksum has changed before the user tries to navigate away.
This is better than relying on 'change' events of form inputs because it takes
into account the fact that users may change the values back to the previous
ones.