Spike "Save and Preview" in a new tab
User research and early user feedback has shown that vast majority of the users
expect this to open in a new tab. Some users found it difficult to navigate
back to the form when we open the preview window in the same tab.

However, to save the draft first we need to submit the form. Unfortunately
buttons do not have a target="_blank" equivalent. After trying a few hackish
solutions I've come to the conclusion that it is best to:
1) Submit the form via Ajax
2.1) If the save succeeds, proceed opening the preview link in a new tab
2.2) If the save fails, prevent the new tab opening and submit the form once
     again, this time without Ajax so that the user can see a validation message

There is a slight downside when the save fails and the user fixes the errors
and tries to "Save and Preview" once more: since the saving happens via Ajax,
validation messages do not disappear.