Send an email notification when a comment is added
When a user comments on a guide edition, notify the edition author and the
author of the current guide edition (may be different when a comment was added
on an old edition) via email.

Emails are currently delivered synchronously even though #deliver_later is used.
This will become asynchronous if we add and configure a job queue.