Do not show the Update type form fields for new guides
User testing has revealed that the 'Update' type box when creating a new guide
is very confusing. Hide this part of the form for new guides and default to
'minor' update type. 'Minor' is the default, because 'Major' requires a change
note to be present and it's not possible to add that if the form is hidden.