Expose latest_editions for topics manager via a helper method
To populate the dropdown fields this data set is required whenever the template
is rendered. Previous code did not take into account that the template is
rendered on create/update failures too. Instead of polluting controller code
with instance variables it's better to expose this as a helper method that
is also lazily evaluated.