Tag guides to topics
We're planning to use information about topics to render breadcrumbs and
subtitles in government-frontend. We also need this for backwards compatibility
with the collections-publisher application (we're planning to migrate topic
management there once its ready).

Known issues:
- If a guide is added to more than one topic, only the latest one will be
recorder. However, we do not expect a guide to belong to more than one topic at
this time.
- We do not have sidekiq or any other queue processor in place, so this will be
executed synchronously and may be slow