Community flavoured guides are selectable for content owner.
For the moment we have decided to use the Guide (along with it's format service_manual_guide) for our service manual community page. Therefore we have a self referential link from Guides to (Community) Guides. In order to distinguish guides and community flavoured guides we decided to put a boolean flag on the guides table.