Allow to prepare for URL slug migrations up-front
Originally we only allowed users to select published guides as a redirect target.
After talking to the users it has became clear that they would prefer doing
the mapping up-front. Once the target guide is published the UI allows users
to click "Migrate" that creates the redirect rule.

This way the laborious part (mapping) can be done before the topics and guides
are published and it needs only a button click later.