Filter guides by content owner using GuideCommunity
We also stop including :content_owner when loading the Guides in the guides index.

The self referential link is causing ambiguity on the tsv column when searching. Removing this will result in an N+1 query. A TODO note has been added and this will hopefully be addressed after we have completed the rush for a deploy.