Switch to "major" for "update_type" when creating redirects.
Content items in publishing-api expect to have an "update_type" of "major"
when they're first created. If it's set to "minor" then publishing-api
fails to create the item, because it requires a "public_updated_at"
field for some reason. For now, this will fix slug migrations
temporarily. Later on, we can think about moving back to "minor", if
need be.