Always index with the format 'service_manual'
It looked like we were sending a format when indexing the root page but not when indexing the guides or topics. Considering we are about to push into the mainstream index with pages outside of the service manual then it makes even more sense to give us a way to distinguish them.