Use the new format types and specify manual in search
We are trying to use the same format types from the content schemas when adding to the search index. We down weight in rummager based on the format.

We need the manual to be "service-manual" because that is how the /service-manual/search page is scoping the search results to the service manual.