Guide#latest_edition_per_edition_group
Returns a collection of editions which are the latest edition per group of editions that share the same version.

This is useful for the guide history page. We want to be able to display a list of all editions (or draft for lack of better naming) without loading all versions for each edition.