Creating a new draft increments the edition version
We have deleted a validation from Edition which stops you from updating details of a published edition. It was causing a problem with the nature of the latest_edition association which was attempting to nullify the previous edition's guide_id foreign key. Removing it doesn't matter because we are no longer changing any editions, only creating new ones.