Stop a previous edition's updated at time being updated
Another result of the has_one #latest_edition when using #build_latest_edition was that the updated_at time was getting set. This is in addition to the guide_id being nullified. So in our hacking to get around the association near the top of GuidesController#update we set updated_at too. The hacking is nasty but work has already started to remove the association and therefore the need for the hack. Phew.