Remove nested attributes in the guide form
 * First pass at replacing the @guide.latest_edition with an @edition attribute

We're trying to remove the nested attributes so any use of #latest_edition= needs to go. For the moment the aim is to attempt to save both a @guide and a related @edition in the same form.

 * Pass error messages to 'shared/form_error_summary' partial

Because our errors are split over two objects it's convenient to be able to combine them as arrays.

 * Detect the unpersisted edition rather than use the association in ApprovalProcess

This was also using the association attribute.

 * Clear the change note and summary when editing a published edition

We lost this functionality when we stopped using Guide#latest_editable_edition in the form.

 * First edition after a published one should have update_type 'major'

This functionality was previously covered by Guide#latest_editable_edition.

 * Remove Guide#latest_editable_edition and Edition#draft_copy

 * Complete guide form nested attributes removal

 * Use partial locals in guide form