Replace Guide#latest_edition association
 * Replace Guide#latest_edition association with an alias to #latest_persisted_edition

The #latest_edition association was causing frustration. It was a has_one association with a scope and it was easy to forget it was a has_one and try and use it as a pure scope. You would call it the first time and the scope would be used to load the latest edition but when you called it a second time it would return the loaded model. We decided to take the time to remove it.

The first step was removing the nested attributes which relied on the association which has been done.

 * Replace use of #latest_edition in factories

 * Replace use of #latest_edition in factory overrides

 * Remove the includes for :latest_edition

They weren't working anyway. Limits are ignored when preloading and don't work with has_ones either https://github.com/rails/rails/issues/10621.

This post talks about a way to get it working using a subquery https://ksylvest.com/posts/2014-12-20/advanced-eager-loading-in-rails

For the moment it's safer to not preload because we are saving so many editions now.

 * Replace use of Edition#build_latest_edition

Now the association is gone the method doesn't exist anymore.

 * Fix a couple of Publisher specs for removal of #latest_edition

 * Reload editions to fix a test for removal of #latest_edition

 * Test fixes for removal of #latest_edition

 * Fix SearchIndexer test for removal of #latest_edition

 * Validate new Guide editions have a content_owner

Previously we could check latest_edition to see if it was valid when it was an attribute. Now we need to search through for the unpersisted edition in the editions association collection.

 * Rename Guide#latest_persisted_edition to #latest_edition