A form object to save guides
 * A very broken GuideForm in the GuideController#new action

A new object to handle some of the logic lurking in the guide controller that handles the saving of two objects in one form.

 * GuideForm#save attempts to assign the forms attributes and save the models

A long way to go still but this mimics the activerecord interface.

 * Validation errors are passed up to GuideForm from models

 * Always reset the state to draft when saving an edition

 * Assign an edition version when saving a GuideForm

 * Test happy path for saving a new GuideForm

 * Pass a GuideForm to the Publisher

We do this because the Publisher needs to be able to react to whether a guide/edition are saved successfully or not and the GuideForm is now the thing doing the saving.

 * Make GuideForm#to_param return the same as the Guide would

 * Load the GuideForm description from the edition

 * Load the GuideForm title from the edition

 * Load GuideForm body from the edition

 * Load GuideForm content_owner_id from edition

 * Assert the guide form slug is correct after saving

 * Set default update_type to be major for GuideForm

 * Successful happy path for saving and updating a GuideForm

 * Reuse @guide_form throughout guides controller

The @guide_form has become the "thing" that the controller juggles. The guide and edition are still accessible from it.

 * Load GuideForm type from the guide

In order to preserve the user's choice to create a guide or guide community we need to instantiate the correct version of the STI.

 * Increment the edition version number in the GuideForm

 * Move defaulting to a major edition after publishing to the GuideForm

 * Move the clearing of the change notes to the GuideForm

 * Remove old strong params from guide controller

 * Remove unused guide reload in publish action