Index the latest published edition in search
Previously we were indexing the latest edition, published or not. This worked when publishing because we were only attempting to index guides when publishing. We are about to add a rake task to reindex all the currently published content and so this SearchIndexer class needed to know which edition was the most recently published.