Do not reindex guides that have been unpublished
Now we support unpublishing we don't want to accidentally add an unpublished guide back into search.

A "live edition" is an edition that is visible on the frontend. So at the moment this is just a published editon. In the future this will also return a withdrawn edition because, unliked unpublished editions, they will still exist on the frontend.

When we get withdrawn editions we should only have to change Edition::STATES_THAT_UPDATE_THE_FRONTEND to include 'withdrawn' and Guide#live_edition should continue to work because it is only the unpublished state we want to exclude from editions we consider as 'live' from editions that update the frontend.