Rake task to republish guides
This currently assumes that every guide has a live edition.

We don't save a new edition or draft for this. It is just republishing the same content and supposedly is only run if there is a change to the GuidePresenter. Our use case this time is that we are changing the frontend and so we need to updated the rendering_app for all content items.