A boolean for whether topics should be visually collapsed
By default it is false. Topics sections are expanded by default. This is considered safer than collapsing by default even though it is more likely it will be collapsed eventually.

The topic has multiple sections which are normally collapsed. Sometimes there aren't enough sections or items within a section to justify to collapsing and it appears there aren't a set of rules that can be applied which will let the frontend decide for itself. Instead we are adding this boolean which the publisher will set so a pair of human eyes can decide whether it should be collapsed or expanded.