Remove strong params from guide controller
At the moment it doesn't matter whether we use strong params or not for the guide controller because we're not doing mass assignment. But if we start we might accidentally let something through. Better to avoid strong params so we will get an error if/when we start mass assigning.