Tag all content to GDS in publishing-api
This app currently tags all content to the GDS in Rummager
(https://github.com/alphagov/service-manual-publisher/blob/f434aa0aa2671
3e829c93b838031dc0abfd2fe25/app/models/guide_search_indexer.rb#L24).

In the near future, Rummager will start indexing the `links` hash from
the publishing-api. This means that we should start sending the
`organisations` tagging to publishing-api too, so that rummager can
copy this.

When we turn on the links-indexing the `organisations` tagging can be
removed from the rummager payload.