Split guide actions between GuideForm and GuideManager
This splits the various actions taken on a guide into two groups.

1) Saving guide content - validating, creating and updating the various fields eg. body, title, description etc.
2) Managing the state of the guide - requesting a review, approving, publishing and discarding

Saving content is handled by the GuideForm. Managing the state is handled by the GuideManager.

This removes more logic from the controller and hopefully the saving vs. managing language makes the guide controller easier to follow.

The previous Publisher object was too optimistic that it was going to be general enough to work for guides and topics. The symptoms of this was that additional functionality started to appear on Publisher that wasn't relevant for topics. eg. Publisher#discard_draft. And when we wanted to extend shared functionality, but only for one type, the controller became a dumping ground. Publisher has now become TopicPublisher as it only deals with Topics.