Remove topic selection validation when creating points
Service standard points do not belong to a topic, so we shouldn't require a topic section id when creating points.

To avoid multiple special cases within GuideForm, we create the PointForm class, and modify the guides controller so that it calls GuideForm.build, which instantiates and returns the relevant type of form object for the thing being created / edited.