Add the points directly to the service standard details
Expanding custom fields isn't quite ready. Here we add the required fields for each point in the service standard to the details hash.

The parent link has been left in the point links payload so when the custom expansion starts working we'll be able to see it in the flesh. For the moment, it is functionless.