Remove the summary from editions
It turned out there was no need to maintain a separate field for the meta description and the short point text.

To avoid changing everything at once we are leaving the summary in the details hash in the schema. This can potentially be removed in favour of using the description field instead but there needs to be a decision on how we distinguish a guide from a point.