Validate presence of title in the model
Although the publishing api currently 'validates' the presence of the title field, validating in the model means that the title field is highlighted correctly, and that all validation messages are displayed together.

Currently if the slug validation fails, the title validation message does not appear because the topic is never pushed to the publishing api.