Do not regenerate the slug if the slug was previously manually edited
We don't persist the edited state, so on page load we try and regenerate the slug based on the title.

If it doesn't match, we can assume the slug was previously manually edited, and so we will not update the slug when the user changes the title etc.

If it matches, the slug was effectively automatically generated previously, and so we will continue to update the slug.