Add a Service Standard Search Indexer
This will be used in a rake task to index the service standard in Rummager. It's very basic and uses the content payload given to us by the presenter, but given how soon the content will start being stored in the publishing api we can revisit this task then to make it nicer.