Fix adding topic sections to lists
The after initialise callback happens to early, before any attributes are actually set on the model. this means that self.topic_id is always nil, so it doesn't correctly set the position.

Instead, make the topic model responsible for adding its own sections – setting the position as it does so.

Unfortunately there are a number of places in the test that manually create! topic sections, so we also need to update them to use the topic section factory.