Don't duplicate change history when republishing guides
During the normal publishing workflow we need to include the change note and reason from the latest edition before it's published, as the publication of a guide does not update its content. So for this case the 'latest' change notes come from the most recent edition, and the previous change history is all published major editions.

However, when republishing the latest live edition will already be a published major edition, so we want to exclude it from the previous change history