Make naming of change history fields consistent
Within the publishing API we call the summary for a change the `change_note` and the 'reason for change' is `reason_for_change`

Confusingly, within our application's edition model and the database schema we call the same fields `change_summary` and `change_note` respectively.

This renames change_summary to `change_note`, and `change_note` to `reason_for_change` to avoid confusion.