Rename TopicsController#respond_for_topic_publication
A little confusingly the 'publication' part was referring to the response from the TopicPublisher but it's easy to think that it's the output from an attempt to publish something. It's just part of the unavoidable wider confusion that publishing a page involves three actions, once of which is publishing.

Called #topic_respond_with to avoid confusion with the Rails method #respond_with which does a similar thing.