The first guide edition has a major update type
The publishing-api requires that the first edition has an update_type of 'major'.

Our content editors had the requirement that they did not want to give a change_note or reason_for_change for the first edition.

In addition to setting the update_type to major we had to change our validations to allow both an empty reason_for_change and to supply a default change_note for the first edition.