Fix guide's topic section validation
The validation wasn't kicking in when creating the first draft because the GuideForm was building an unpersisted an `topic_section_guides` which then wasn't considered `empty?`. We fix this by only building the association if a topic_section_id is supplied.

This implementation will cause a problem if we one day start allowing the user to assign a null topic section. But for the moment, that seems unlikely.