Add an Email Alert Signup Presenter
In order to allow users to subscribe to email alerts for topics within the service manual, we need to publish an email alert signup. This will be rendered by the email-alert-frontend and provide an interface for the user to sign up for alerts.

This is the first step towards that, creating a presenter that represents the email alert signup we'll eventually send to the publishing api.

The `subscriber_list` item in the details hash 'defines' what emails will be sent to the list. When new or updated guides going through the publishing api 'match' against the links, email alerts will be triggered and sent to the people signed up.