Remove the guide tagger job
Guides used to be associated with topics by editing the topic, so we used to need to update the guides whenever a topic was saved.

However topics are now associated when creating and editing guides, so the association can be made there.

This means that A) email alerts will work correctly and B) the breadcrumb for a guide will be correct without having to go and edit the topic.