Add a task to republish all topics
We will need to republish all topics to add the link to their respective email alert signup pages.

There is currently no easy way to do this, other than to manually go to each topic page within publisher, save a draft, and republish it. This automates this process.

Note that because we do not currently have a concept of editions for topics within the publisher, running this script will publish the latest version of every topic, regardless of whether they’ve previously been published or not.