Add checkbox for whether topics appear on homepage
We need to exclude the communities topic from the topics on the homepage. Rather than hard code this, this makes it possible to specify whether a topic should appear on the homepage or not.

Once the homepage is deployed we can configure which topics appear on the homepage appropriately to meet needs.