Link topics to the homepage if included
By using the parent relationship the topics will automatically be updated and appear as children of the homepage, without having to republish the homepage every time we add or update a topic.

We could configure our own bidirectional expansion in the publishing api, but this keep things simple until such a time when we have e.g. multiple types of children.