Fix publish service standard rake task
Since e0891e9, the service standard no longer receives a set of points when instantiated.

The rake task was not updated as part of this change, which resulted in an ArgumentError exception being thrown when the rake task was run.

This commit removes the redundant argument, which allows the rake task to be run again.