Move introduction into description
The service manual frontend now uses description instead of details/introduction (since 925610d) so ensure we set a description. This is currently broken on frontend master with no description being shown.

Additionally, within the service standard content schema this introduction already appears as description and there is no `introduction` within the details hash, so this aligns what is published with what is in the schema.