Add task to remove orphaned associations
The associations between Guides and TopicSectionGuides was not previously marked as dependent.

This meant that when unpublished guides had their draft discarded, the topic_section_guide row remained but was associated with a Guide that does not exist.

This breaks the topic page within publisher as it tries to loop over all guides and print their title, but the non-existent guide is nil.