Only consider published guides when building list of related communities
When building the list of content owners, which is used to create the list of related communities for the topic page, we currently consider guides that may not have been published yet.

This can be confusing if the topic gets republished before the guide gets published, as the community will appear on the topic page before it actually gets associated through the publishing of the guide.