Block creating a draft edition with a published version
It shouldn't be possible to edit a published version. The version number should be incremented.

This validation is being added as a result of this happening in production. It was possible to create a draft for a published version by having a guide in a 'ready' state open in two tabs. Clicking 'Publish' in one tab and then clicking 'Save' (to make a new draft) in the other tab would create a draft without incrementing the version number.