Only allow changes to a guide if it hasn't changed since opening
We encountered a bug where a guide was published while it was currently being edited by someone else. This resulted in an edition version getting published twice, which we want to avoid.

We have defended against this here but not allowing a guide to be altered if the guide has been changed since it was opened for editing in the browser. The unique bit of information that identifies the last edition, or fingerprint, is just the edition ID because editions are (in practise) immutable. We only create new editions rather than update them.

This should defend against other unexpected changes in state..